14.Usunięto plik klasy która nie musi być publiczna.Dodano jej do innego pliku
