+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -35,5 +35,40 @@ public class IntLinkedList {
         last = last.getPrev();
         return ret;
     }
+    
+    private class Node {
+        private int value;
+        private Node prev;
+    	private Node next;
+
+        public Node(int valueTemp) {
+            setValue(valueTemp);
+        }
+
+    	public Node getNext() {
+    		return next;
+    	}
+
+    	public void setNext(Node next) {
+    		this.next = next;
+    	}
+
+    	public Node getPrev() {
+    		return prev;
+    	}
+
+    	public void setPrev(Node prev) {
+    		this.prev = prev;
+    	}
+
+    	public int getValue() {
+    		return value;
+    	}
+
+    	public void setValue(int value) {
+    		this.value = value;
+    	}
+
+    }
 
 }
